QuantumDynamics

Documentation for QuantumDynamics.

QuantumDynamics.ModelsOpSum.ising

QuantumDynamics.ModelsPS.X

QuantumDynamics.ModelsPS.XXX_ts

QuantumDynamics.ModelsPS.XXZ_ts

QuantumDynamics.ModelsPS.XX_ts

QuantumDynamics.ModelsPS.XYZ_ts

QuantumDynamics.ModelsPS.XY_ts

QuantumDynamics.ModelsPS.ising_ts

QuantumDynamics.ModelsTEBD.ising

QuantumDynamics.heisenberg

QuantumDynamics.heisenberg

QuantumDynamics.heisenberg — Method

heisenberg(H::Matrix, O::Matrix, observables, dt::Real, tmax::Real)

Heisenberg evolution with dense matrices. Evolve the operator O in the Heisenberg picture under the Hamiltonian H for a time tmax with time step dt. observables is a list of functions, which are applied to the operator O at each time step.

QuantumDynamics.heisenberg — Method

heisenberg(H::Vector{ITensor}, O::MPO, observables, dt::Real, tmax::Real; bd=100, cutoff=1E-8)

Heisenberg evolution with MPO (TEBD). Evolve the operator O in the Heisenberg picture under the Hamiltonian H for a time tmax with time step dt. observables is a list of functions, which are applied to the operator O at each time step. H is a list of ITensors representing the Hamiltonian terms.

QuantumDynamics.ModelsPS.X — Method

X(N), Y(N), Z(N)

Total X, Y, Z operators for a system of size N.

QuantumDynamics.ModelsPS.XXX_ts — Method

XXX_ts(N)

1D XXX model $H= \sum_{j=1}^N ( X_j X_{j+1} + Y_j Y_{j+1} + Z_j Z_{j+1})$ Returns a OperatorTS1D object.

QuantumDynamics.ModelsPS.XXZ_ts — Method

XXZ_ts(N, delta)

1D XXZ model $H= \sum_{j=1}^N ( X_j X_{j+1} + Y_j Y_{j+1} + \Delta Z_j Z_{j+1})$ Returns a OperatorTS1D object.

QuantumDynamics.ModelsPS.XX_ts — Method

XX_ts(N)

1D XX model $H= \sum_{j=1}^N ( X_j X_{j+1} + Y_j Y_{j+1} )$ Returns a OperatorTS1D object.

QuantumDynamics.ModelsPS.XYZ_ts — Method

XYZ_ts(N, deltay, deltaz)

1D XYZ model $H= \sum_{j=1}^N ( X_j X_{j+1} + \Delta_Y Y_j Y_{j+1} + \Delta_Z Z_j Z_{j+1})$ Returns a OperatorTS1D object.

QuantumDynamics.ModelsPS.XY_ts — Method

XY_ts(N, delta)

1D XY model $H= \sum_{j=1}^N ( X_j X_{j+1} + \Delta Y_j Y_{j+1} )$ Returns a OperatorTS1D object.

QuantumDynamics.ModelsPS.ising_ts — Method

ising_ts(N, hx, hz)

Hamiltonian for the 1D Ising model with periodic boundary conditions. $H= \sum_{j=1}^N ( Z_j Z_{j+1} + h_x X_j + h_z Z_j)$ Returns a OperatorTS1D object.

QuantumDynamics.ModelsOpSum.ising — Method

ising(N, hx, hz)

Hamiltonian for the 1D Ising model with periodic boundary conditions. $H= \sum_{j=1}^N ( Z_j Z_{j+1} + h_x X_j + h_z Z_j)$

QuantumDynamics.ModelsTEBD.ising — Method

ising(N, hx, hz, sites)

Hamiltonian for the 1D Ising model with periodic boundary conditions. $H= \sum_{j=1}^N ( Z_j Z_{j+1} + h_x X_j + h_z Z_j)$