Blip decomposition of the path integral

The blip decomposition of the path integral, as described in Blip1 and Blip2, is a different representation of the path integral that significantly reduces the number of "paths" that need to be considered.

API

QuantumDynamics.Blip.build_augmented_propagator — Function

build_augmented_propagator(; fbU::Matrix{ComplexF64}, Jw::Vector{T}, β::Real, dt::Real, ntimes::Int, extraargs::BlipArgs=BlipArgs(), svec=[1.0 -1.0], reference_prop=false, verbose::Bool=false) where {T<:SpectralDensities.SpectralDensity}

Builds the propagators, augmented with the influence of the harmonic baths defined by the spectral densities Jw, upto ntimes time-steps without iteration using the blip decomposition. The paths are, consequently, generated in the space of unique blips and not stored. So, while the space requirement is minimal and constant, the time complexity for each time-step grows by an additional factor of $b$, where $b$ is the number of unique blip-values. The i^th bath, described by Jw[i], interacts with the system through the diagonal operator with the values of svec[j,:].

QuantumDynamics.Blip.build_augmented_propagator_QuAPI_TTM — Function

build_augmented_propagator_QuAPI_TTM(; fbU::Matrix{ComplexF64}, Jw::Vector{T}, β::Real, dt::Real, ntimes::Int, extraargs::BlipArgs=BlipArgs(), svec=[1.0 -1.0], reference_prop=false, verbose::Bool=false) where {T<:SpectralDensities.SpectralDensity}

Builds the propagators, augmented with the influence of the harmonic baths defined by the spectral densities Jw, upto ntimes time-steps without iteration using the blip decomposition. The paths are, consequently, generated in the space of unique blips and not stored. So, while the space requirement is minimal and constant, the time complexity for each time-step grows by an additional factor of $b$, where $b$ is the number of unique blip-values. The i^th bath, described by Jw[i], interacts with the system through the diagonal operator with the values of svec[j,:]. In this version, multiple ``types'' of propagators are calculated. These are required to make the TTM scheme consistent with QuAPI splitting.

QuantumDynamics.Blip.BlipArgs — Type

Filtration parameters for blips. Currently has the maximum number of blips allowed which by default is -1 (implying all blips are allowed).