Bloch-Redfield Master Equation

API

QuantumDynamics.BlochRedfield.propagate — Function

propagate(; Hamiltonian::AbstractMatrix{ComplexF64}, Jw::AbstractVector{T}, β::Real, ρ0::AbstractMatrix{ComplexF64}, dt::Real, ntimes::Int, sys_ops::Vector{Matrix{ComplexF64}}, extraargs::Utilities.DiffEqArgs=Utilities.DiffEqArgs()) where {T<:SpectralDensities.AnalyticalSpectralDensity}

Given a system Hamiltonian, the spectral densities describing the solvent, Jw, and an inverse temperature, this uses Bloch-Redfield Master Equations to propagate the input initial reduced density matrix, ρ0, with a time-step of dt for ntimes time steps. The ith bath, described by Jw[i], interacts with the system through the operator with the values of svec[j]. The default solver used here is Tsit5 with a relative and absolute error cutoffs of 1e-10.

QuantumDynamics.BlochRedfield.get_Rtensor — Function

get_Rtensor(eigvals, eigvecs, Jw::AbstractVector{T}, svec::AbstractVector{Matrix{ComplexF64}}, β::Real) where {T<:SpectralDensities.AnalyticalSpectralDensity}

Calculates the Bloch-Redfield R tensor given the eigenvalues, eigvals, and eigenvectors, eigvecs, of the system Hamiltonian, an inverse temperature β, and a number of baths specified by their spectral densities, Jw, and the operator through which they interact, svec.