Complex-time Path Integrals: Quantum Correlation Functions

API

QuantumDynamics.ComplexTimePI.A_of_t — Function

A_of_t(; Hamiltonian::AbstractMatrix{ComplexF64}, β::Real, t::Real, N::Int, Jw::AbstractVector{<:SpectralDensities.SpectralDensity}, svec::AbstractMatrix{<:Real}, A, B, extraargs::Utilities.TensorNetworkArgs=Utilities.TensorNetworkArgs())

Calculates $Tr_{env}(U(t) exp(-β H/2) A exp(-β H/2) U^{-1}(t))$ for a system interacting with an environment at a time-point t using the tensor network path integral method. This can be used for thermodynamics or for calculating correlation functions.

Relevant references:

Topaler, M.; Makri, N. Quantum rates for a double well coupled to a dissipative bath: Accurate path integral results and comparison with approximate theories. The Journal of Chemical Physics 1994, 101 (9), 7500-7519.

Bose, A. Quantum correlation functions through tensor network path integral. The Journal of Chemical Physics 2023, 159 (21), 214110.

Arguments:

Hamiltonian: system Hamiltonian

Jw: array of spectral densities

svec: diagonal elements of system operators through which the corresponding baths interact. QuAPI currently only works for baths with diagonal coupling to the system.

β: inverse temperature

t: time at which the function is evaluated

N: number of path integral discretizations

A: system operator to be evaluated

extraargs: extra arguments for the tensor network algorithm. Contains the cutoff threshold for SVD filtration, the maximum bond dimension, maxdim, and the algorithm of applying an MPO to an MPS.

QuantumDynamics.ComplexTimePI.correlation_function_tnpi — Function

correlation_function_tnpi(; Hamiltonian::AbstractMatrix{ComplexF64}, β::Real, tfinal::Real, dt::Real, N::Int, Jw::AbstractVector{<:SpectralDensities.SpectralDensity}, svec::AbstractMatrix{<:Real}, A, B, Z::Real, extraargs::Utilities.TensorNetworkArgs=Utilities.TensorNetworkArgs(), verbose::Bool=false, output::Union{Nothing,HDF5.Group}=nothing)

Calculates the $<A(0) B(t)> / Z$ correlation function for a system interacting with an environment upto a maximum time of tfinal with a time-step of dt using the tensor network path integral method.

Relevant references:

Topaler, M.; Makri, N. Quantum rates for a double well coupled to a dissipative bath: Accurate path integral results and comparison with approximate theories. The Journal of Chemical Physics 1994, 101 (9), 7500-7519.

Bose, A. Quantum correlation functions through tensor network path integral. The Journal of Chemical Physics 2023, 159 (21), 214110.

Arguments:

Hamiltonian: system Hamiltonian

Jw: array of spectral densities

svec: diagonal elements of system operators through which the corresponding baths interact. QuAPI currently only works for baths with diagonal coupling to the system.

β: inverse temperature

tfinal: maximum time till which the correlation function is evaluated

dt: time-step for evaluating correlation function

N: number of path integral discretizations

A: system operator evaluated at time zero

B: array of system operators evaluated at time t

Z: partition function for normalization

extraargs: extra arguments for the tensor network algorithm. Contains the cutoff threshold for SVD filtration, the maximum bond dimension, maxdim, and the algorithm of applying an MPO to an MPS.

verbose: verbosity

output: output HDF5 file for storage of results

QuantumDynamics.ComplexTimePI.unnormalized_correlation_function_tnpi — Function

unnormalized_correlation_function_tnpi(; Hamiltonian::AbstractMatrix{ComplexF64}, β::Real, t::Real, N::Int, Jw::AbstractVector{<:SpectralDensities.SpectralDensity}, svec::AbstractMatrix{<:Real}, A, B, extraargs::Utilities.TensorNetworkArgs=Utilities.TensorNetworkArgs())

Calculates the $<A(0) B(t)>$ correlation function for a system interacting with an environment at a time-point t using the tensor network path integral method.

Relevant references:

Topaler, M.; Makri, N. Quantum rates for a double well coupled to a dissipative bath: Accurate path integral results and comparison with approximate theories. The Journal of Chemical Physics 1994, 101 (9), 7500-7519.

Bose, A. Quantum correlation functions through tensor network path integral. The Journal of Chemical Physics 2023, 159 (21), 214110.

Arguments:

Hamiltonian: system Hamiltonian

Jw: array of spectral densities

svec: diagonal elements of system operators through which the corresponding baths interact. QuAPI currently only works for baths with diagonal coupling to the system.

β: inverse temperature

t: time at which the correlation function is evaluated

N: number of path integral discretizations

A: system operator evaluated at time zero

B: array of system operators evaluated at time t

extraargs: extra arguments for the tensor network algorithm. Contains the cutoff threshold for SVD filtration, the maximum bond dimension, maxdim, and the algorithm of applying an MPO to an MPS.

QuantumDynamics.ComplexTimePI.unnormalized_correlation_function_quapi — Function

unnormalized_correlation_function_quapi(; Hamiltonian::AbstractMatrix{ComplexF64}, β::Real, t::Real, N::Int, Jw::AbstractVector{<:SpectralDensities.SpectralDensity}, svec::AbstractMatrix{<:Real}, A, B, extraargs::QuAPI.QuAPIArgs=QuAPI.QuAPIArgs())

Calculates the $<A(0) B(t)>$ correlation function for a system interacting with an environment at a time-point t using QuAPI.

Relevant references:

Topaler, M.; Makri, N. Quantum rates for a double well coupled to a dissipative bath: Accurate path integral results and comparison with approximate theories. The Journal of Chemical Physics 1994, 101 (9), 7500-7519.

Arguments:

Hamiltonian: system Hamiltonian

Jw: array of spectral densities

svec: diagonal elements of system operators through which the corresponding baths interact. QuAPI currently only works for baths with diagonal coupling to the system.

β: inverse temperature

t: time at which the correlation function is evaluated

N: number of path integral discretizations

A: system operator evaluated at time zero

B: system operator evaluated at time t

extraargs: extra arguments for the QuAPI algorithm. Contains a cutoff threshold for filtration of paths.