Quantum Correlation Functions using Path Integrals

API

QuantumDynamics.ComplexTNPI.A_of_t — Function

A_of_t(; Hamiltonian::AbstractMatrix{ComplexF64}, β::Real, t::Real, N::Int, Jw::AbstractVector{<:SpectralDensities.SpectralDensity}, svec::AbstractMatrix{<:Real}, A, B, extraargs::Utilities.TensorNetworkArgs=Utilities.TensorNetworkArgs())

Calculates $Tr_{env}(U(t) exp(-β H/2) A exp(-β H/2) U^{-1}(t))$ for a system interacting with an environment at a time-point t using the tensor network path integral method. This can be used for thermodynamics or for calculating correlation functions.

Arguments:

Hamiltonian: system Hamiltonian

Jw: array of spectral densities

svec: diagonal elements of system operators through which the corresponding baths interact. QuAPI currently only works for baths with diagonal coupling to the system.

β: inverse temperature

t: time at which the function is evaluated

N: number of path integral discretizations

A: system operator to be evaluated

extraargs: extra arguments for the tensor network algorithm. Contains the cutoff threshold for SVD filtration, the maximum bond dimension, maxdim, and the algorithm of applying an MPO to an MPS.

QuantumDynamics.ComplexTNPI.complex_correlation_function — Function

complex_correlation_function(; Hamiltonian::AbstractMatrix{ComplexF64}, β::Real, tfinal::Real, dt::Real, N::Int, Jw::AbstractVector{<:SpectralDensities.SpectralDensity}, svec::AbstractMatrix{<:Real}, A, B, Z::Real, extraargs::Utilities.TensorNetworkArgs=Utilities.TensorNetworkArgs(), verbose::Bool=false, output::Union{Nothing,HDF5.Group}=nothing, exec=ThreadedEx())

Calculates the $<A(0) B(t_c)> / Z$ correlation function for a system interacting with an environment upto a maximum time of tfinal with a time-step of dt using the tensor network path integral method.

Arguments:

Hamiltonian: system Hamiltonian

Jw: array of spectral densities

svec: diagonal elements of system operators through which the corresponding baths interact. QuAPI currently only works for baths with diagonal coupling to the system.

β: inverse temperature

tfinal: maximum time till which the correlation function is evaluated

dt: time-step for evaluating correlation function

N: number of path integral discretizations

A: system operator evaluated at time zero

B: array of system operators evaluated at time t

Z: partition function for normalization

extraargs: extra arguments for the tensor network algorithm. Contains the cutoff threshold for SVD filtration, the maximum bond dimension, maxdim, and the algorithm of applying an MPO to an MPS.

verbose: verbosity

output: output HDF5 file for storage of results

exec: FLoops.jl execution policy

QuantumDynamics.ComplexQuAPI.A_of_t — Function

A_of_t(; Hamiltonian::AbstractMatrix{ComplexF64}, β::Float64, t::Float64, N::Int64, Jw::AbstractVector{<:SpectralDensities.SpectralDensity}, svec::AbstractMatrix{Float64}, A, extraargs::QuAPI.QuAPIArgs=QuAPI.QuAPIArgs(), exec=FLoops.ThreadedEx())

Calculates $Tr_{env}(U(t) exp(-β H/2) A exp(-β H/2) U^{-1}(t))$ for a system interacting with an environment at a time-point t using the tensor network path integral method. This can be used for thermodynamics or for calculating correlation functions.

Arguments:

Hamiltonian: system Hamiltonian

Jw: array of spectral densities

svec: diagonal elements of system operators through which the corresponding baths interact. QuAPI currently only works for baths with diagonal coupling to the system.

β: inverse temperature

t: time at which the function is evaluated

N: number of path integral discretizations

A: system operator to be evaluated

extraargs: extra arguments for the tensor network algorithm. Contains the cutoff threshold for SVD filtration, the maximum bond dimension, maxdim, and the algorithm of applying an MPO to an MPS.

exec: FLoops.jl execution policy

QuantumDynamics.ComplexQuAPI.complex_correlation_function — Function

complex_correlation_function(; Hamiltonian::AbstractMatrix{ComplexF64}, β::Real, tfinal::Real, dt::Real, N::Int, Jw::AbstractVector{<:SpectralDensities.SpectralDensity}, svec::AbstractMatrix{<:Real}, A, B, Z::Real, extraargs::QuAPI.QuAPIArgs=QuAPI.QuAPIArgs(), verbose::Bool=false, output::Union{Nothing,HDF5.Group}=nothing, exec=ThreadedEx())

Calculates the $<B(t_c) A(0)> / Z$ correlation function for a system interacting with an environment upto a maximum time of tfinal with a time-step of dt using the tensor network path integral method.

Arguments:

Hamiltonian: system Hamiltonian

Jw: array of spectral densities

svec: diagonal elements of system operators through which the corresponding baths interact. QuAPI currently only works for baths with diagonal coupling to the system.

β: inverse temperature

tfinal: maximum time till which the correlation function is evaluated

dt: time-step for evaluating correlation function

N: number of path integral discretizations

A: system operator evaluated at time zero

B: system operator evaluated at time t

Z: partition function for normalization

extraargs: extra arguments for the tensor network algorithm. Contains the cutoff threshold for SVD filtration, the maximum bond dimension, maxdim, and the algorithm of applying an MPO to an MPS.

verbose: verbosity

output: output HDF5 file for storage of results

exec: FLoops.jl execution policy

QuantumDynamics.ComplexQuAPI.adaptive_kink_A_of_t — Function

adaptive_kink_A_of_t(; Hamiltonian::AbstractMatrix{ComplexF64}, β::Float64, t::Float64, N::Int64, Jw::AbstractVector{<:SpectralDensities.SpectralDensity}, svec::AbstractMatrix{Float64}, A, extraargs::QuAPI.QuAPIArgs=QuAPI.QuAPIArgs(), exec=FLoops.ThreadedEx())

Calculates $Tr_{env}(U(t) exp(-β H/2) A exp(-β H/2) U^{-1}(t))$ for a system interacting with an environment at a time-point t using the tensor network path integral method. This can be used for thermodynamics or for calculating correlation functions.

Arguments:

Hamiltonian: system Hamiltonian

Jw: array of spectral densities

svec: diagonal elements of system operators through which the corresponding baths interact. QuAPI currently only works for baths with diagonal coupling to the system.

β: inverse temperature

t: time at which the function is evaluated

N: number of path integral discretizations

A: system operator to be evaluated

extraargs: extra arguments for the tensor network algorithm. Contains the cutoff threshold for SVD filtration, the maximum bond dimension, maxdim, and the algorithm of applying an MPO to an MPS.

exec: FLoops.jl execution policy

QuantumDynamics.ComplexQuAPI.adaptive_kink_complex_correlation_function — Function

adaptive_kink_complex_correlation_function(; Hamiltonian::AbstractMatrix{ComplexF64}, β::Real, tfinal::Real, dt::Real, N::Int, Jw::AbstractVector{<:SpectralDensities.SpectralDensity}, svec::AbstractMatrix{<:Real}, A, B, Z::Real, extraargs::QuAPI.QuAPIArgs=QuAPI.QuAPIArgs(), verbose::Bool=false, output::Union{Nothing,HDF5.Group}=nothing, exec=ThreadedEx())

Calculates the $<B(t_c) A(0)> / Z$ correlation function for a system interacting with an environment upto a maximum time of tfinal with a time-step of dt using the tensor network path integral method.

Arguments:

Hamiltonian: system Hamiltonian

Jw: array of spectral densities

svec: diagonal elements of system operators through which the corresponding baths interact. QuAPI currently only works for baths with diagonal coupling to the system.

β: inverse temperature

tfinal: maximum time till which the correlation function is evaluated

dt: time-step for evaluating correlation function

N: number of path integral discretizations

A: system operator evaluated at time zero

B: system operator evaluated at time t

Z: partition function for normalization

extraargs: extra arguments for the tensor network algorithm. Contains the cutoff threshold for SVD filtration, the maximum bond dimension, maxdim, and the algorithm of applying an MPO to an MPS.

verbose: verbosity

output: output HDF5 file for storage of results

exec: FLoops.jl execution policy