Generalized Quantum Master Equation (GQME)

This module enables usage of the Nakajima-Zwanzig master equation in presence of empirical losses described by Lindblad equations. The memory kernel is computed from the path integral augmented propagators through the transfer tensors.

API

QuantumDynamics.GQME.propagate — Function

propagate(; Hamiltonian::AbstractMatrix{<:Complex}, Jw::Vector{T}, β::Real, ρ0::AbstractMatrix{<:Complex}, dt::Real, ntimes::Int, rmax::Int, kmax::Union{Int,Nothing}=nothing, path_integral_routine, extraargs::Utilities.ExtraArgs, svec=[1.0 -1.0], reference_prop=false, verbose::Bool=false, L::Union{Nothing,Vector{Matrix{ComplexF64}}}=nothing, output::Union{Nothing,HDF5.Group}=nothing) where {T<:SpectralDensities.SpectralDensity}

Propagate an initial density matrix ρ0 under a Hamiltonian, environments described by the spectral densities Jw held at an inverse temperature β. The memory kernel is obtained using the transfer tensor method.

QuantumDynamics.GQME.propagate_with_memory_kernel — Function

propagate_with_memory_kernel(; K::AbstractArray{<:Complex,3}, fbU::AbstractMatrix{<:Complex}, ρ0::AbstractMatrix{<:Complex}, dt::Real, ntimes::Int, L::Union{Nothing,Vector{Matrix{ComplexF64}}}=nothing)

Given a memory kernel K, and the bare forward-backward propagator fbU, propagate the initial density matrix ρ0. Can additionally involve empirical losses through Lindblad jump operators L.