Pairwise-Connected Tensor Network Path Integral

It has been shown in PC-TNPI that a general tensor network can be formulated that manifestly captures the symmetries present in the path integral expression. It can be used by itself or as a starting point for deriving other tensor network decompositions of path integrals.

API

QuantumDynamics.PCTNPI.build_augmented_propagator — Function

build_augmented_propagator(; fbU::Matrix{ComplexF64}, Jw::Vector{T}, β::Real, dt::Real, ntimes::Int, svec=[1.0 -1.0], reference_prop=false, verbose::Bool=false) where {T<:SpectralDensities.SpectralDensity}

Builds the propagators, augmented with the influence of the harmonic baths defined by the spectral densities Jw, upto ntimes time-steps using the PCTNPI approach. The paths are, consequently, generated in the space of unique blips and not stored. So, while the space requirement is minimal and constant, the time complexity for each time-step grows by an additional factor of $b$, where $b$ is the number of unique blip-values. The i^th bath, described by Jw[i], interacts with the system through the diagonal operator with the values of svec[j,:].