Propagators

API

QuantumDynamics.Propagators.calculate_bare_propagators — Function

calculate_bare_propagators(; Hamiltonian::AbstractMatrix{<:Complex}, dt::AbstractFloat, ntimes=1, external_fields::Union{Nothing,Vector{Utilities.ExternalField}}=nothing)

This function calculates the bare propagators for a given Hamiltonian and under the influence of the external_fields with a time-step of dt for ntimes time steps.

QuantumDynamics.Propagators.ReferenceChoice — Type

ReferenceChoice{choice}

This empty struct holds the particular choice of algorithm for selecting reference forces. choice currently can be one of fixed or ehrenfest.