Spectral Densities

The interaction of a quantum system with a condensed phase environment is often captured through the spectral density. QuantumDynamics has a built-in support for a few of the most common spectral densities and allows for easy incorporation of other spectral densities.

API

QuantumDynamics.SpectralDensities — Module

Collection of spectral densities commonly used to describe solvents.

QuantumDynamics.SpectralDensities.ExponentialCutoff — Type

ExponentialCutoff(; ξ, ωc, n=1.0, Δs=2.0)

Construct a model spectral density with an exponential cutoff.

$J(ω) = \frac{2π}{Δs^2} ξ \frac{ω^n}{ω_c^{n-1}} \exp\left(-\frac{ω}{ωc}\right)$

where Δs is the distance between the two system states. The model is Ohmic if n = 1, sub-Ohmic if n < 1, and super-Ohmic if n > 1.

QuantumDynamics.SpectralDensities.DrudeLorentz — Type

DrudeLorentz(; λ, γ, Δs=2.0)

Construct a model spectral density with a Drude-Lorentz cutoff.

$J(ω) = \frac{2λ}{Δs^2} \frac{ω γ}{ω^2 + γ^2}$

where Δs is the distance between the two system states.

QuantumDynamics.SpectralDensities.matsubara_decomposition — Function

matsubara_decomposition(sd::DrudeLorentz, num_modes::Int, β::Float64)

Returns the Matsubara frequencies, γ, and the expansion coefficients, c.

QuantumDynamics.SpectralDensities.SpectralDensityTable — Type

SpectralDensityTable <: ContinuousSpectralDensity

Spectral density provided in tabular form. Contains a vector of ωs and a vector corresponding to jws.

QuantumDynamics.SpectralDensities.tabulate — Function

tabulate(sd::T, full_real::Bool=true, npoints::Int=10000) where {T<:AnalyticalSpectralDensity}

Returns a table with ω and J(ω) for ω between -ωmax to ωmax if full_real is true. Otherwise the table ranges for ω between 0 and ωmax with npoints.

tabulate(sd::SpectralDensityTable, full_real::Bool=true, npoints::Int=100001)

Returns sd.ω and sd.jw.