Utilities

Collection of some utilities for simulations.

API

QuantumDynamics.Utilities.ExternalField — Type

ExternalField provides an abstract interface for encoding an external field, V(t), interacting with the system through the operator, coupling_op.

QuantumDynamics.Utilities.unhash_path — Function

unhash_path(path_num::Int, ntimes::Int, sdim::Int)

Construct a path for a system with sdim dimensions, corresponding to the number path_num, with ntimes time steps.

QuantumDynamics.Utilities.apply_propagator — Function

apply_propagator(; propagators, ρ0, ntimes, dt)

Apply a series of ntimes propagators to an initial reduced density matrix ρ0 and return the result as a tuple of (time, ρs).

Many of the algorithms require extra, method-specific arguments. These are implemented as subtypes of Utilities.ExtraArgs.

QuantumDynamics.Utilities.ExtraArgs — Type

Abstract type for encoding all the method specific numerical parameters.

QuantumDynamics.Utilities.DiffEqArgs — Type

Extra parameters for solving differential equations. Currently has a threshold for magnitude-based filtering. The default values are: reltol = 1e-10 abstol = 1e-10 solver = Tsit5()