η-Coefficients

The η-coefficients are discretizations of the bath correlation function required for simulations using the QuAPI influence functionals. QuantumDynamics provides facilities for generating these coefficients and storing them in a way that utilizes the limited time-translational symmetry that they demonstrate. These coefficients have been listed in many papers, the first being QuAPI1.

API

QuantumDynamics.EtaCoefficients.EtaCoeffs — Type

EtaCoefficients holds the various discretized η-coefficients required for a QuAPI-based simulation. These are the minimum number of coefficients required, stored using time-translational symmetry wherever possible.

The values are stored as follows:

η00: The self-interaction of the two terminal time points.

ηmm: The self-interaction of all intermediate points.

η0m: The interaction between a terminal and an intermediate point at different time separations.

ηmn: The interaction between two intermediate points at different time separations.

η0e: The interaction between the two terminal points at different time separations.

QuantumDynamics.EtaCoefficients.calculate_η — Function

calculate_η(specdens::SpectralDensities.ContinuousSpectralDensity; β::Real, dt::Real, kmax::Int, imaginary_only=false)

Calculates the η-coefficients from an analytic spectral density and returns them as an object of the structure EtaCoeffs. The integrations involved are done using trapezoidal integration

calculate_η(specdens::SpectralDensity.DiscreteOscillators; β::Real, dt::Real, kmax::Int, classical::Bool=false, imaginary_only=false)

Calculates the η-coefficients from a discretized set of harmonic modes and returns them as an object of the structure EtaCoeffs. The integrations involved are converted to sums over frequency modes.