Pairwise-Connected Tensor Network Path Integral

It has been shown in PC-TNPI that a general tensor network can be formulated that manifestly captures the symmetries present in the path integral expression. It can be used by itself or as a starting point for deriving other tensor network decompositions of path integrals.

API

QuantumDynamics.PCTNPI.build_augmented_propagator — Function

build_augmented_propagator(; fbU::AbstractArray{ComplexF64,3}, Jw::AbstractVector{T}, β::Real, dt::Real, ntimes::Int, kmax::Union{Int,Nothing}=nothing, svec=[1.0 -1.0], reference_prop=false, verbose::Bool=false, extraargs::PCTNPIArgs=PCTNPIArgs(), output::Union{Nothing,HDF5.Group}=nothing, from_TTM::Bool=false, exec=ThreadedEx()) where {T<:SpectralDensities.SpectralDensity}

Builds the propagators, augmented with the influence of the harmonic baths defined by the spectral densities Jw, upto ntimes time-steps using the PCTNPI approach. The j^th bath, described by Jw[j], interacts with the system through the diagonal operator with the values of svec[j,:].

Relevant references:

Bose, A. Pairwise Connected Tensor Network Representation of Path Integrals. Physical Review B 2022, 105 (2), 024309. https://doi.org/10.1103/PhysRevB.105.024309.

Arguments:

fbU: time-series of forward-backward propagators

Jw: array of spectral densities

svec: diagonal elements of system operators through which the corresponding baths interact. QuAPI currently only works for baths with diagonal coupling to the system.

β: inverse temperature

dt: time-step for recording the density matrices

ntimes: number of time steps of simulation

extraargs: extra arguments for the PCTNPI algorithm. Contains the cutoff threshold for SVD filtration, the maximum bond dimension, maxdim, and the algorithm of applying an MPO to an MPS.

exec: FLoops.jl execution policy