Utilities

Collection of some utilities for simulations.

API

QuantumDynamics.Utilities.get_BLAS_implementation — Function

get_BLAS_implementation()

Reports the BLAS implementation under use. The default implementation used by Julia is OpenBlas. MKL is supported through the external package MKL.jl, which needs to be installed and loaded before the loading of QuantumDynamics.jl

QuantumDynamics.Utilities.ExternalField — Type

ExternalField provides an abstract interface for encoding an external field, V(t), interacting with the system through the operator, coupling_op.

QuantumDynamics.Utilities.unhash_path — Function

unhash_path(path_num::Int, ntimes::Int, sdim::Int)

Construct a path for a system with sdim dimensions, corresponding to the number path_num, with ntimes time steps.

QuantumDynamics.Utilities.unhash_path_blips — Function

unhash_path_blips(ntimes::Int, sdim::Int, nblips::Int)

Construct all the paths for a system with sdim dimensions with ntimes time steps and nblips blips.

QuantumDynamics.Utilities.apply_propagator — Function

apply_propagator(; propagators, ρ0, ntimes, dt)

Apply a series of ntimes propagators to an initial reduced density matrix ρ0 and return the result as a tuple of (time, ρs).

QuantumDynamics.Utilities.convert_ITensor_to_matrix — Function

convert_ITensor_to_matrix(tens, sinit, sterm)

Converts an ITensor with two indices to a matrix. The index sinit is mapped to the column and sterm is mapped to the row.

There are a few utilities for creating specific kinds of Hamiltonians.

QuantumDynamics.Utilities.create_tls_hamiltonian — Function

create_tls_hamiltonian(; ϵ, Δ)

Creates a two-level system Hamiltonian:

$H = \frac{ϵ}{2}σ_z - \frac{Δ}{2}σ_x$

QuantumDynamics.Utilities.create_nn_hamiltonian — Function

create_nn_hamiltonian(; site_energies, couplings, periodic::Bool)

Creates a nearest neighbour Hamiltonian with the given site_energies and couplings. Periodic boundary conditions can also be used by passing true into the periodic argument.

Many of the algorithms require extra, method-specific arguments. These are implemented as subtypes of Utilities.ExtraArgs.

QuantumDynamics.Utilities.ExtraArgs — Type

Abstract type for encoding all the method specific numerical parameters.

QuantumDynamics.Utilities.DiffEqArgs — Type

Extra parameters for solving differential equations. Currently has a threshold for magnitude-based filtering. The default values are: reltol = 1e-10 abstol = 1e-10 solver = Tsit5()