Utilities

Collection of some utilities for simulations.

API

ITensor Utilities

QuantumDynamics.Utilities.convert_ITensor_to_matrix — Function

convert_ITensor_to_matrix(tens, sinit, sterm)

Converts an ITensor with two indices to a matrix. The index sinit is mapped to the column and sterm is mapped to the row.

QuantumDynamics.Utilities.identity_MPO — Function

identity_MPO(sites)

Returns the identity MPO based on the given sites.

QuantumDynamics.Utilities.MPO_to_MPS — Function

MPO_to_MPS(ρ::MPO, fbcombiner)

Convert a given MPO to an MPS by combining the two site indices on every tensor into a single one using the vector of combiner tensors.

QuantumDynamics.Utilities.MPS_to_MPO — Function

MPO_to_MPS(ρ::MPO, fbcombiner)

Split a given MPS to an MPO by using the vector of combiner tensors.

ITensors.expect — Function

ITensors.expect(ρ::MPO, ops; kwargs...)

Extends ITensors' expect function to handle density matrices in the form of MPOs.

Generic Utilities

QuantumDynamics.Utilities.get_BLAS_implementation — Function

get_BLAS_implementation()

Reports the BLAS implementation under use. The default implementation used by Julia is OpenBlas. MKL is supported through the external package MKL.jl, which needs to be installed and loaded before the loading of QuantumDynamics.jl

QuantumDynamics.Utilities.trapezoid — Function

trapezoid(x, y; discrete::Bool=false)

Returns the trapezoidal integration of y with respect to x. If discrete is set to true, then returns sum of y.

QuantumDynamics.Utilities.commutator — Function

commutator(A, B)

Returns the commutator A and B: AB - BA.

QuantumDynamics.Utilities.calculate_Liouvillian — Function

calculate_Liouvillian(Hamiltonian::AbstractMatrix{ComplexF64})

Returns the Liouvillian corresponding to the given Hamiltonian.

calculate_Liouvillian(H::OpSum, sites)

Returns the forward-backward space combiner and the Liouvillian MPO corresponding to the Hamiltonian provided as an ITensor OpSum, to be built on the sites.

QuantumDynamics.Utilities.ExternalField — Type

ExternalField provides an abstract interface for encoding an external field, V(t), interacting with the system through the operator, coupling_op.

QuantumDynamics.Utilities.hash_path — Function

hash_path(states, sdim)

Returns the hashed location of a path for a system with sdim dimensions.

QuantumDynamics.Utilities.unhash_path — Function

unhash_path(path_num::Int, ntimes::Int, sdim::Int)

Construct a path for a system with sdim dimensions, corresponding to the number path_num, with ntimes time steps.

QuantumDynamics.Utilities.unhash_path_blips — Function

unhash_path_blips(ntimes::Int, sdim::Int, nblips::Int)

Construct all the paths for a system with sdim dimensions with ntimes time steps and nblips blips.

QuantumDynamics.Utilities.apply_propagator — Function

apply_propagator(; propagators, ρ0, ntimes, dt)

Apply a series of ntimes propagators to an initial reduced density matrix ρ0 and return the result as a tuple of (time, ρs).

There are a few utilities for creating specific kinds of Hamiltonians.

QuantumDynamics.Utilities.create_tls_hamiltonian — Function

create_tls_hamiltonian(; ϵ, Δ)

Creates a two-level system Hamiltonian:

$H = \frac{ϵ}{2}σ_z - \frac{Δ}{2}σ_x$

QuantumDynamics.Utilities.create_nn_hamiltonian — Function

create_nn_hamiltonian(; site_energies, couplings, periodic::Bool)

Creates a nearest neighbour Hamiltonian with the given site_energies and couplings. Periodic boundary conditions can also be used by passing true into the periodic argument.

Many of the algorithms require extra, method-specific arguments. These are implemented as subtypes of Utilities.ExtraArgs.

QuantumDynamics.Utilities.ExtraArgs — Type

Abstract type for encoding all the method specific numerical parameters.

QuantumDynamics.Utilities.DiffEqArgs — Type

Extra parameters for solving differential equations. Currently has a threshold for magnitude-based filtering. The default values are: reltol = 1e-10 abstol = 1e-10 solver = Tsit5()